fix(docs): restructure sidebar and localization handling
