+++ apps/docs/src/components/MarkdownContent.astro
@@ -1,4 +1,4 @@
-<div class="markdown-content"><slot /></div>
+<div class="markdown-content sl-markdown-content"><slot /></div>
 
 <script>
   document.addEventListener('DOMContentLoaded', function () {
