+++ apps/docs/src/components/SidebarSublist.astro
@@ -1,6 +1,6 @@
 ---
 import { isActiveOrParentPath } from '../utils/navigation'
-import { NavigationCategory } from '../content/config'
+import { NavigationCategory } from '../sidebar-config'
 import arrowUpRightIcon from '@assets/sidebar/arrow-up-right.svg?raw'
 
 // Add TypeScript interface declaration for Window
