+++ apps/docs/src/components/TableOfContents.astro
@@ -1,7 +1,7 @@
 ---
 import TableOfContentsList from './TableOfContentsList.astro'
 
-const { toc } = Astro.props
+const { toc } = Astro.locals.starlightRoute
 ---
 
 {
