+++ apps/docs/src/sidebar-config.ts
@@ -1,29 +1,6 @@
-import { docsSchema, i18nSchema } from '@astrojs/starlight/schema'
-import { defineCollection, z } from 'astro:content'
-import { generateI18nSchema } from 'src/i18n/generateI18nSchema'
 import { localizePath } from 'src/i18n/utils'
 
-import type { NavigationGroup } from '../utils/navigation'
-
-export const collections = {
-  docs: defineCollection({
-    schema: docsSchema({
-      extend: z.object({
-        licence: z.string().optional(),
-        distribution: z.string().optional(),
-        hideTitleOnPage: z.boolean().optional(),
-        llmUrl: z.string().optional(),
-        toolboxLlmUrl: z.string().optional(),
-      }),
-    }),
-  }),
-  i18n: defineCollection({
-    type: 'data',
-    schema: i18nSchema({
-      extend: generateI18nSchema(),
-    }),
-  }),
-}
+import type { NavigationGroup } from './utils/navigation'
 
 export enum NavigationCategory {
   MAIN,
